+++ src/zephyr/client.ts
@@ -14,6 +14,7 @@ import { GetProjects } from "./tool/project/get-projects";
 import { GetStatuses } from "./tool/status/get-statuses";
 import { CreateTestCaseIssueLink } from "./tool/test-case/create-issue-link";
 import { CreateTestCase } from "./tool/test-case/create-test-case";
+import { CreateTestScript } from "./tool/test-case/create-test-script";
 import { CreateTestCaseWebLink } from "./tool/test-case/create-web-link.ts";
 import { GetTestCase } from "./tool/test-case/get-test-case";
 import { GetTestCases } from "./tool/test-case/get-test-cases";
@@ -90,6 +91,7 @@ export class ZephyrClient implements Client {
       new CreateTestCaseWebLink(this),
       new CreateTestCaseIssueLink(this),
       new CreateFolder(this),
+      new CreateTestScript(this),
     ];
 
     for (const tool of tools) {